Improved Custom Algorithm
Modified the custom algorithm to now sort the pictures by the ratio of the cost of each picture to the amount of space it takes up. Also fixed a problem that caused the backtracking to not work properly (as the wallLength was not updated properly when a picture was removed). Also added a new test file with 5000 pictures.