Updated Custom Algorithm
Updated the custom algorithm to improve speed. It now terminates when it finds a solution that uses all of the wall space, but if the backup solution has a better value, then it will use the backup solution (but still terminate). This prevents it from iterating over the many smaller subsets that will not have a greater value than the backup (greatly improving the overall speed of the algorithm)